feat(shell): integrate topnav content provider and update header
- Wrap WorkspaceLayout with TopNavProvider in workspace layout
- Refactor Header to register topnav content instead of rendering UI
- Remove unused state and UI logic from Header component

This change sets up a persistent top navigation bar infrastructure for the app shell.
